Crate utils_lib

utils-lib

A bunch of utilities that I uses across several project. Instead of copying the same code everywhere I finally decided to put everything into a library. I will probably continue grow this library.

I won’t maintain this crate actively and I might introduce breaking change at anytime. I also might be slow to merge the changes on the main branch. Also I won’t publish this crate on crates.io so you will have to specify the git url in the cargo.toml file.

[dependencies]
utils-lib = { version = "0.1.0", git = "https://git.noxie.ch/ABouttefeux/utils-lib" }

If you want the bleeding edge feature you might want to target the develop branch using

[dependencies]
utils-lib = { git = "https://git.noxie.ch/ABouttefeux/utils-lib", branch = "develop" }

again it might break at anytime so you might want to target an specific commit

[dependencies]
utils-lib = { git = "https://git.noxie.ch/ABouttefeux/utils-lib", branch = "develop", rev="<commit hash>" }

Re-exports

pub use self::coordinate::Axis2D;

pub use self::coordinate::Coordinate;

pub use self::number::abs_diff;

pub use self::number::PositiveFloat;

pub use self::number::ValidationGuard;

pub use self::number::ZeroOneBoundedFloat;

Modules

coordinate

Module containing Coordinate a 2d coordinate and Axis2D an enumeration of the x and y axis.

error

Contains the errors definitions.

number

Contains number and math utilities.

Macros

impl_op_trait

impl std::ops trait for a given type wrapper with a given method that gives or is derefed into the wrapped value. The traits have to be imported by you at the place of invocation, (see example).

trait_sealed

Creates a trait Sealed into a private module private.

Derive Macros

Getter

Derive getter macro

Sealed

Derive the Sealed trait