Module utils_lib::coordinate

Module containing Coordinate a 2d coordinate and Axis2D an enumeration of the x and y axis.

Structs

Coordinate

A two dimensional vector.

CoordinateIterator

Iterator on a coordinate Coordinate. It is the type return by Coordinate::into_iter (and Coordinate::iter and Coordinate::iter_mut thought behind implicit type) .

Enums

Axis2D

Represent the Axis in 2 dimensions. It can be either in the x direction i.e. Self::Vertical or the y direction, i.e. Self::Horizontal.