Enum utils_lib::coordinate::Axis2D

pub enum Axis2D {
    Vertical,
    Horizontal,
}

Represent the Axis in 2 dimensions. It can be either in the x direction i.e. Self::Vertical or the y direction, i.e. Self::Horizontal.

Variants

Vertical

X axis

Horizontal

Y Axis

Implementations

impl Axis2D

pub const AXIS: [Self; 2] = _

All the possible axis

pub const fn from_index(index: usize) -> Option<Self>

Convert an index into an Axis2D

Example

use utils_lib::coordinate::Axis2D;

assert_eq!(Axis2D::from_index(0), Some(Axis2D::Vertical));
assert_eq!(Axis2D::from_index(1), Some(Axis2D::Horizontal));
assert_eq!(Axis2D::from_index(2), None);
assert_eq!(Axis2D::from_index(3), None);
//...

pub const fn to_index(self) -> usize

Convert an Axis2D into an index

Example

use utils_lib::coordinate::Axis2D;

assert_eq!(Axis2D::Vertical.to_index(), 0);
assert_eq!(Axis2D::Horizontal.to_index(), 1);

pub const fn as_index(self) -> &'static usize

Convert an Axis2D as an index

Example

use utils_lib::coordinate::Axis2D;

assert_eq!(Axis2D::Vertical.as_index(), &0);
assert_eq!(Axis2D::Horizontal.as_index(), &1);

pub const fn perpendicular(self) -> Self

Get the perpendicular axis

Example

use utils_lib::coordinate::Axis2D;

assert_eq!(Axis2D::Vertical.perpendicular(), Axis2D::Horizontal);
assert_eq!(Axis2D::Horizontal.perpendicular(), Axis2D::Vertical);

pub const fn coordinate_usize(self) -> Coordinate<usize>

Convert an Axis2D into a cardinal direction in the form of a Coordinate::<usize>

Example

use utils_lib::coordinate::{Axis2D, Coordinate};

assert_eq!(Axis2D::Vertical.coordinate_usize(), Coordinate::new(1, 0));
assert_eq!(Axis2D::Horizontal.coordinate_usize(), Coordinate::new(0, 1));