Struct utils_lib::coordinate::Coordinate

pub struct Coordinate<T> {
    pub x: T,
    pub y: T,
}

A two dimensional vector.

Fields
x: T

the x coordinate

y: T

the y coordinate

Implementations

impl<T> Coordinate<T>

pub const fn new(x: T, y: T) -> Self

Create a new Coordinate with two values for, respectively, the x and y coordinate.

pub const fn x(&self) -> &T

Get the x coordinate.

pub fn x_mut(&mut self) -> &mut T

Get a mut reference on the x coordinate.

pub const fn y(&self) -> &T

Get the y coordinate.

pub fn y_mut(&mut self) -> &mut T

Get a mut reference on the y coordinate.

pub const fn get(&self, axis: Axis2D) -> &T

Get the coordinate given by the Axis2D direction.

pub fn get_mut(&mut self, axis: Axis2D) -> &mut T

Get a mutable reference on the coordinate given by the Axis2D direction.

pub fn iter( &self ) -> impl Iterator<Item = &T> + DoubleEndedIterator + FusedIterator + ExactSizeIterator

Get an iterator on the coordinate elements

pub fn iter_mut( &mut self ) -> impl Iterator<Item = &mut T> + DoubleEndedIterator + FusedIterator + ExactSizeIterator

Get an iterator on the coordinate elements as mutable reference

pub const fn as_tuple(&self) -> (&T, &T)

Get the Coordinate as a tuple references

pub fn as_tuple_mut(&mut self) -> (&mut T, &mut T)

Get the Coordinate as a tuple mut references

pub const fn as_array(&self) -> [&T; 2]

Get the Coordinate as an array references

pub fn as_array_mut(&mut self) -> [&mut T; 2]

Get the Coordinate as an array mut references

pub const fn as_ref(&self) -> Coordinate<&T>

Get the Coordinate as a Coordinate references

pub fn as_mut(&mut self) -> Coordinate<&mut T>

Get the Coordinate as a Coordinate mut references

impl<T> Coordinate<T>

Some “move” conversion function

pub fn into_tuple(self) -> (T, T)

Get the Coordinate as a tuple

pub fn into_array(self) -> [T; 2]

Get the Coordinate as an array

impl<T: Copy> Coordinate<T>

Const conversion function using Copy as a bound on T.

pub const fn into_tuple_const(self) -> (T, T)

Get the Coordinate as a tuple. This is a const function.

pub const fn into_array_const(self) -> [T; 2]

Get the Coordinate as an array. This is a const function.

impl<'a, T> Coordinate<T>where T: PartialOrd, &'a T: Sub + 'a, <&'a T as Sub>::Output: Add,

pub fn s1_distance( &'a self, other: &'a Self ) -> <<&'a T as Sub>::Output as Add>::Output

Manhattan distances

Example

use utils_lib::coordinate::Coordinate;

let coord_zero = Coordinate::new(0_i32, 0_i32);
assert_eq!(coord_zero.s1_distance(&coord_zero), 0_i32);

let coord = Coordinate::new(0_i32, 1_i32);
assert_eq!(coord.s1_distance(&coord_zero), 1_i32);

let coord = Coordinate::new(1_i32, 0_i32);
assert_eq!(coord.s1_distance(&coord_zero), 1_i32);

let coord = Coordinate::new(3_i32, 4_i32);
assert_eq!(coord.s1_distance(&coord_zero), 7_i32);

let coord_1 = Coordinate::new(10_i32, 22_i32);
let coord_2 = Coordinate::new(13_i32, 21_i32);
assert_eq!(coord_1.s1_distance(&coord_2), 4_i32);
assert_eq!(coord_2.s1_distance(&coord_1), 4_i32);