Derive Macro utils_lib::Sealed

#[derive(Sealed)]

Derive the Sealed trait

Panic

panic if the derive macro is not applied to an struct, enum or union

Example

use utils_lib_derive::{trait_sealed, Sealed};

// this create a module named [`private`] with a trait named [`Sealed`]
// without method inside that module.
trait_sealed!();

#[derive(Sealed)]
struct S;

// this trait is sealed and cannot me implemented outside of this crate
// because [`Sealed`] is a private trait that can't be implemented outside
// of this crate.
pub trait Trait: private::Sealed {}

impl Trait for S {}