Macro utils_lib::impl_op_trait

macro_rules! impl_op_trait {
    ($s:ty, $method:ident, Add) => { ... };
    ($s:ty, $method:ident, Mul) => { ... };
    ($s:ty, $method:ident, Div) => { ... };
    ($s:ty, $method:ident, Sub) => { ... };
    ($s:ty, $method:ident, Rem) => { ... };
    ($s:ty, $method:ident, $t1:ident, $f1:ident, $t2:ident, $f2:ident) => { ... };
}

impl std::ops trait for a given type wrapper with a given method that gives or is derefed into the wrapped value. The traits have to be imported by you at the place of invocation, (see example).

Example

// this has to be imported for the macro to work.
use std::ops::{Add, AddAssign, Sub, SubAssign};

use utils_lib::impl_op_trait;
use utils_lib_derive::Getter;

#[derive(Debug, Clone, Copy, PartialEq, Getter)]
struct Wrapper {
    #[get(Const)]
    #[get_mut(Pub)] // this create a method `float_mut` that gives `&mut f64`
    float: f64,
}

impl std::ops::Deref for Wrapper {
    type Target = f64;

    fn deref(&self) -> &Self::Target {
        self.float()
    }
}

impl_op_trait!(Wrapper, float_mut, Add);
impl_op_trait!(Wrapper, float_mut, Sub);

let w1 = Wrapper { float: 1_f64 };
let w2 = Wrapper { float: 2_f64 };

let w3 = Wrapper { float: 3_f64 };

assert_eq!(w1 + w2, w3);
assert_eq!(w1 + &w2, w3);
assert_eq!(&w1 + w2, w3);
assert_eq!(&w1 + &w2, w3);
assert_eq!(w2 - w1, w1);

let mut w2 = w2;
w2 += w1;

assert_eq!(w2, w3);

another possibility is that instead of using a direct mut getter we can use a struct similar to crate::number::ValidationGuard.

// this has to be imported for the macro to work.
use std::ops::{Add, AddAssign, Sub, SubAssign};
use std::ops::{Deref, DerefMut};

use utils_lib::impl_op_trait;
use utils_lib_derive::Getter;

// We need Copy for the macro to work
#[derive(Debug, Clone, Copy, PartialEq, Getter)]
struct Wrapper {
    #[get(Const)]
    float: f64,
}

// we need this trait for the macro to work
impl Deref for Wrapper {
    type Target = f64;

    fn deref(&self) -> &Self::Target {
        self.float()
    }
}

#[derive(Debug, Getter)]
pub struct ValidationGuard<'a> {
    reference: &'a mut Wrapper,
    #[get(Const)]
    #[get_mut]
    float: f64,
}

impl<'a> Deref for ValidationGuard<'a> {
    type Target = f64;

    fn deref(&self) -> &Self::Target {
        self.float()
    }
}

// we need this trait for the macro to work
impl<'a> DerefMut for ValidationGuard<'a> {
    fn deref_mut(&mut self) -> &mut Self::Target {
        self.float_mut()
    }
}

// when we drop the guard we change the value in the wrapper with the tracked value
// we could add some validation here
impl<'a> Drop for ValidationGuard<'a> {
    fn drop(&mut self) {
        self.reference.float = self.float;
    }
}

impl Wrapper {
    pub fn float_mut(&'_ mut self) -> ValidationGuard<'_> {
        ValidationGuard {
            float: self.float,
            reference: self,
        }
    }
}

impl_op_trait!(Wrapper, float_mut, Add);
impl_op_trait!(Wrapper, float_mut, Sub);

let w1 = Wrapper { float: 1_f64 };
let w2 = Wrapper { float: 2_f64 };

let w3 = Wrapper { float: 3_f64 };

assert_eq!(w1 + w2, w3);
assert_eq!(w1 + &w2, w3);
assert_eq!(&w1 + w2, w3);
assert_eq!(&w1 + &w2, w3);
assert_eq!(w2 - w1, w1);

let mut w2 = w2;
w2 += &w1;
assert_eq!(w2, w3);