Module utils_lib::number

Contains number and math utilities.

Re-exports

pub use self::positive_float::ConversionError as PositiveFloatConversionError;

pub use self::positive_float::PositiveFloat;

pub use self::sign::Sign;

pub use self::zero_one_bounded_float::ConversionError as ZeroOneBoundedFloatConversionError;

pub use self::zero_one_bounded_float::ZeroOneBoundedFloat;

Modules

positive_float

Contains PositiveFloat.

sign

Contains the definition of Sign and related notions.

zero_one_bounded_float

Contains ZeroOneBoundedFloat.

Structs

ValidationGuard

A structure created by PositiveFloat::float_mut, it can be DerefMut as an &mut f64. It ensure data validation on Drop. If the data is not valid it is set to 0.

Traits

Validation

Trait for type that have some validation step for data

Functions

abs_diff

Do the absolute difference of two numbers. In mathematical notation it is |a-b|.

gcd

Find the greater common divider

lcm

Find the lowest common multiplier