Struct utils_lib::number::positive_float::PositiveFloat

pub struct PositiveFloat(/* private fields */);

A float that is >= 0 and is not f64::NAN or f64::INFINITY.

Implementations

impl PositiveFloat

pub const ZERO: Self = _

Value 0

pub const ONE: Self = _

Value 1

pub const MAX: Self = _

Maximum value

pub fn new(float: f64) -> Result<Self, ConversionError>

Create a new Self from a f64. It returns Some only if the float is valid (Self::validate_data), i.e. it is >= 0 it is not f64::NAN and not f64::INFINITY.

Errors

If float is smaller than zero it returns ConversionError::TooLow.

If float is f64::INFINITY it returns ConversionError::Infinity.

If float is f64::NAN it returns ConversionError::Nan.

Example

use utils_lib::number::PositiveFloatConversionError;
use utils_lib::PositiveFloat;

PositiveFloat::new(0_f64)?;
PositiveFloat::new(2.5_f64)?;
PositiveFloat::new(6.7E10_f64)?;

assert_eq!(
    PositiveFloat::new(f64::INFINITY),
    Err(PositiveFloatConversionError::Infinity)
);
assert_eq!(
    PositiveFloat::new(-f64::INFINITY),
    Err(PositiveFloatConversionError::TooLow)
);
assert_eq!(
    PositiveFloat::new(-f64::NAN),
    Err(PositiveFloatConversionError::Nan)
);
assert_eq!(
    PositiveFloat::new(-1_f64),
    Err(PositiveFloatConversionError::TooLow)
);
assert_eq!(
    PositiveFloat::new(-100_f64),
    Err(PositiveFloatConversionError::TooLow)
);

pub fn new_or_default(float: f64) -> Self

Create a new Self with the float as value if it is valid ( >= 0 finite and not f64::NAN) or return the default value (0) instead.

Example

use utils_lib::PositiveFloat;

assert_eq!(
    PositiveFloat::new_or_default(0_f64),
    PositiveFloat::new(0_f64)?
);
assert_eq!(
    PositiveFloat::new_or_default(2.5_f64),
    PositiveFloat::new(2.5_f64)?
);
assert_eq!(
    PositiveFloat::new_or_default(6.7E10_f64),
    PositiveFloat::new(6.7E10_f64)?
);

assert_eq!(
    PositiveFloat::new_or_default(f64::INFINITY),
    PositiveFloat::ZERO
);
assert_eq!(
    PositiveFloat::new_or_default(-f64::INFINITY),
    PositiveFloat::default()
);
assert_eq!(
    PositiveFloat::new_or_default(-f64::NAN),
    PositiveFloat::default()
);
assert_eq!(
    PositiveFloat::new_or_default(-1_f64),
    PositiveFloat::default()
);
assert_eq!(
    PositiveFloat::new_or_default(-100_f64),
    PositiveFloat::default()
);

pub fn new_or_bounded(float: f64) -> Self

or return 0 or to f64::MAX if the value is infinity instead.

Note that contrary to Self::new_or_default when given infinity it gives bac Self::MAX

use utils_lib::PositiveFloat;

assert_eq!(PositiveFloat::new_or_bounded(0.5_f64).float(), 0.5_f64);
assert_eq!(PositiveFloat::new_or_bounded(-0.5_f64), PositiveFloat::ZERO);
assert_eq!(PositiveFloat::new_or_bounded(f64::NAN), PositiveFloat::ZERO);
assert_eq!(
    PositiveFloat::new_or_bounded(f64::INFINITY),
    PositiveFloat::MAX
);
// PositiveFloat::new_or_default instead return 0
assert_eq!(
    PositiveFloat::new_or_default(f64::INFINITY),
    PositiveFloat::ZERO
);

pub const fn float(self) -> f64

Get the underling float. It could also be accessed by using Deref, note that std::ops::DerefMut is not implemented.

pub fn float_mut(&mut self) -> ValidationGuard<'_, Self>

Returns a way to mutate the underlying float. If the final value is not valid, It is set to 0 or to f64::MAX if the value is infinity. See ValidationGuard.

pub fn checked_sub(self, other: Self) -> Result<Self, ConversionError>

Returns the value of the subtraction of two numbers if it doesn’t underflow. It works in the same spirit as usize::checked_sub.

Errors

See Self::new

Example

use utils_lib::PositiveFloat;

let p1 = PositiveFloat::new(1_f64)?;
let p2 = PositiveFloat::new(2_f64)?;

assert_eq!(
    p1.checked_sub(p2),
    Err(PositiveFloatConversionError::TooLow)
);
assert_eq!(p2.checked_sub(p1), Ok(PositiveFloat::new(1_f64)?));

pub fn saturating_sub(self, other: Self) -> Self

Do the subtraction of two PositiveFloat saturating at 0. It works in the same spirit as usize::saturating_sub

Example

use utils_lib::PositiveFloat;

let p1 = PositiveFloat::new(1_f64)?;
let p2 = PositiveFloat::new(2_f64)?;

assert_eq!(p1.saturating_sub(p2), PositiveFloat::new(0_f64)?);
assert_eq!(p2.saturating_sub(p1), PositiveFloat::new(1_f64)?);