Struct utils_lib::number::zero_one_bounded_float::ZeroOneBoundedFloat

pub struct ZeroOneBoundedFloat(/* private fields */);

A float that f is 0 <= f <= 1 and is not NaN.

Implementations

impl ZeroOneBoundedFloat

pub const ZERO: Self = _

Value 0

pub const ONE: Self = _

Value 1

pub fn new(float: f64) -> Result<Self, ConversionError>

Create a new Self from a f64. It returns Some only if the float is valid (Self::validate_data), i.e. it is >= 0 and <= 1.

Errors

If float is smaller than zero it returns ConversionError::TooLow.

If float is bigger than one it returns ConversionError::TooBig.

If float is f64::NAN it returns ConversionError::Nan.

Example

use utils_lib::{number::zero_one_bounded_float::ConversionError, ZeroOneBoundedFloat};

assert!(ZeroOneBoundedFloat::new(0_f64).is_ok());
assert!(ZeroOneBoundedFloat::new(0.001_f64).is_ok());
assert!(ZeroOneBoundedFloat::new(0.6_f64).is_ok());
assert!(ZeroOneBoundedFloat::new(1_f64).is_ok());

assert_eq!(
    ZeroOneBoundedFloat::new(2.5_f64),
    Err(ConversionError::TooBig)
);
assert_eq!(
    ZeroOneBoundedFloat::new(6.7E10_f64),
    Err(ConversionError::TooBig)
);

assert_eq!(
    ZeroOneBoundedFloat::new(f64::INFINITY),
    Err(ConversionError::TooBig)
);
assert_eq!(
    ZeroOneBoundedFloat::new(-f64::INFINITY),
    Err(ConversionError::TooLow)
);
assert_eq!(
    ZeroOneBoundedFloat::new(-f64::NAN),
    Err(ConversionError::Nan)
);
assert_eq!(
    ZeroOneBoundedFloat::new(-1_f64),
    Err(ConversionError::TooLow)
);
assert_eq!(
    ZeroOneBoundedFloat::new(-100_f64),
    Err(ConversionError::TooLow)
);

pub fn new_or_default(float: f64) -> Self

Create a new Self with the float as value if it is valid ( >= 0 and <= 1) or return the default value (0) instead.

Example

use utils_lib::ZeroOneBoundedFloat;

assert_eq!(
    ZeroOneBoundedFloat::new_or_default(0_f64),
    ZeroOneBoundedFloat::new(0_f64)?
);
assert_eq!(
    ZeroOneBoundedFloat::new_or_default(0.5_f64),
    ZeroOneBoundedFloat::new(0.5_f64)?
);
assert_eq!(
    ZeroOneBoundedFloat::new_or_default(1_f64),
    ZeroOneBoundedFloat::new(1_f64)?
);

assert_eq!(
    ZeroOneBoundedFloat::new_or_default(f64::INFINITY),
    ZeroOneBoundedFloat::default()
);
assert_eq!(
    ZeroOneBoundedFloat::new_or_default(-1_f64),
    ZeroOneBoundedFloat::default()
);

pub fn new_or_bounded(float: f64) -> Self

or return 0 for value < 0 and 1 for value > 1. If the value is f64::NAN it use the value 0.

Note that contrary to Self::new_or_default if values are > 1 it creates Self::ONE

use utils_lib::ZeroOneBoundedFloat;

assert_eq!(
    ZeroOneBoundedFloat::new_or_default(1.5_f64),
    ZeroOneBoundedFloat::ZERO
);
assert_eq!(
    ZeroOneBoundedFloat::new_or_bounded(1.5_f64),
    ZeroOneBoundedFloat::ONE
);
assert_eq!(
    ZeroOneBoundedFloat::new_or_bounded(f64::NAN),
    ZeroOneBoundedFloat::ZERO
);
assert_eq!(
    ZeroOneBoundedFloat::new_or_bounded(0.5_f64).float(),
    0.5_f64
);

pub const fn float(self) -> f64

Get the underling float. It could also be accessed by using Deref, note that std::ops::DerefMut is not implemented.

pub fn float_mut(&mut self) -> ValidationGuard<'_, Self>

Returns a way to mutate the underlying float. If the final value is not valid, It is set to 0. See ValidationGuard.

pub fn checked_sub(self, other: Self) -> Result<Self, ConversionError>

Returns the value of the subtraction of two numbers if it doesn’t underflow. It works in the same spirit as usize::checked_sub.

Errors

See Self::new

Example

use utils_lib::ZeroOneBoundedFloat;

let p1 = ZeroOneBoundedFloat::new(0.3_f64)?;
let p2 = ZeroOneBoundedFloat::new(0.6_f64)?;

assert_eq!(
    p1.checked_sub(p2),
    Err(ZeroOneBoundedFloatConversionError::TooLow)
);
assert_eq!(p2.checked_sub(p1), Ok(ZeroOneBoundedFloat::new(0.3_f64)?));

pub fn saturating_sub(self, other: Self) -> Self

Do the subtraction of two ZeroOneBoundedFloat saturating at 0.

Example

TODO

use utils_lib::ZeroOneBoundedFloat;

let p1 = ZeroOneBoundedFloat::new(0.3_f64).unwrap();
let p2 = ZeroOneBoundedFloat::new(0.6_f64).unwrap();

assert_eq!(
    p1.saturating_sub(p2),
    ZeroOneBoundedFloat::new(0_f64).unwrap()
);
assert_eq!(
    p2.saturating_sub(p1),
    ZeroOneBoundedFloat::new(0.3_f64).unwrap()
);

pub fn checked_add(self, other: Self) -> Result<Self, ConversionError>

Returns the value of the addition of two numbers if it doesn’t overflow. It works in the same spirit as Self::checked_sub but with the upper bound.

Errors

See Self::new

Example

use utils_lib::number::ZeroOneBoundedFloatConversionError;
use utils_lib::ZeroOneBoundedFloat;

let p1 = ZeroOneBoundedFloat::new(0.5_f64)?;
let p2 = ZeroOneBoundedFloat::new(0.4_f64)?;
let p3 = ZeroOneBoundedFloat::new(0.6_f64)?;

assert_eq!(p1.checked_add(p2), Ok(ZeroOneBoundedFloat::new(0.9_f64)?));

assert_eq!(
    p1.checked_add(p3),
    Err(ZeroOneBoundedFloatConversionError::TooBig)
);

pub fn saturating_add(self, other: Self) -> Self

Do the addition of two ZeroOneBoundedFloat saturating at 1. It works in the same spirit as Self::saturating_sub but with the upper bound.

Example

use utils_lib::ZeroOneBoundedFloat;

let p1 = ZeroOneBoundedFloat::new(0.5_f64)?;
let p2 = ZeroOneBoundedFloat::new(0.4_f64)?;
let p3 = ZeroOneBoundedFloat::new(0.6_f64)?;

assert_eq!(p1.saturating_add(p2), ZeroOneBoundedFloat::new(0.9_f64)?);

assert_eq!(p1.saturating_add(p3), ZeroOneBoundedFloat::ONE);

pub fn checked_div(self, other: Self) -> Result<Self, ConversionError>

Returns the value of the division of two numbers if it doesn’t overflow. It works in the same spirit as Self::checked_add.

Errors

See Self::new

Example

use utils_lib::number::ZeroOneBoundedFloatConversionError;
use utils_lib::ZeroOneBoundedFloat;

let p1 = ZeroOneBoundedFloat::new(0.1_f64)?;
let p2 = ZeroOneBoundedFloat::new(0.5_f64)?;

assert_eq!(p1.checked_div(p2), Ok(ZeroOneBoundedFloat::new(0.2_f64)?));

assert_eq!(
    p2.checked_div(p1),
    Err(ZeroOneBoundedFloatConversionError::TooBig)
);

assert_eq!(
    p1.checked_div(ZeroOneBoundedFloat::ZERO),
    Err(ZeroOneBoundedFloatConversionError::TooBig)
);

pub fn saturating_div(self, other: Self) -> Self

Do the division of two ZeroOneBoundedFloat saturating at 1. It works in the same spirit as Self::saturating_add.

Example

use utils_lib::ZeroOneBoundedFloat;

let p1 = ZeroOneBoundedFloat::new(0.1_f64)?;
let p2 = ZeroOneBoundedFloat::new(0.5_f64)?;

assert_eq!(p1.saturating_div(p2), ZeroOneBoundedFloat::new(0.2_f64)?);

assert_eq!(p2.saturating_div(p1), ZeroOneBoundedFloat::ONE);

assert_eq!(
    p1.saturating_div(ZeroOneBoundedFloat::ZERO),
    ZeroOneBoundedFloat::ONE
);